# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hebrew RTL E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: brandId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: inStock
          in: query
          schema:
            type: ["boolean", "null"]
        - name: sort
          in: query
          schema:
            type: ["string", "null"]
            enum: [price_asc, price_desc, newest, popular, rating, null]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: tags
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Product list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/homepage products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/top-selling:
    get:
      operationId: getTopSellingProducts
      tags: [products]
      summary: Get top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top selling products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/slug/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get product by SEO slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── BRANDS ────────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/loyalty:
    patch:
      operationId: adjustUserLoyalty
      tags: [users]
      summary: Adjust user loyalty points (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustLoyaltyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── CART ──────────────────────────────────────────────────────────────────
...
  /recently-viewed:
    get:
      operationId: getRecentlyViewed
      tags: [recentlyViewed]
      summary: Get recently viewed products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recently viewed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: trackProductView
      tags: [recentlyViewed]
      summary: Track a product view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackViewBody"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── LOYALTY ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success

    # ── Category ──
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nameHe:
          type: string
        nameEn:
          type: ["string", "null"]
        slug:
          type: string
        descriptionHe:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        costPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        specs:
          type: object
          additionalProperties: true
        stockQuantity:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        viewsCount:
          type: integer
        salesCount:
          type: integer
        ratingAverage:
          type: number
        ratingCount:
          type: integer
        weight:
          type: ["number", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nameHe
        - slug
        - price
        - stockQuantity
        - isActive
        - isFeatured
        - viewsCount
        - salesCount
        - ratingAverage
        - ratingCount
        - images
        - tags
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        nameHe:
          type: string
        nameEn:
          type: ["string", "null"]
        slug:
          type: string
        descriptionHe:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        costPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        specs:
          type: object
          additionalProperties: true
        stockQuantity:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        weight:
          type: ["number", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
      required:
        - nameHe
        - slug
        - price
        - stockQuantity
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages

    # ── User / Auth ──
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, admin]
        loyaltyPoints:
          type: integer
        loyaltyTier:
          type: string
          enum: [bronze, silver, gold, vip]
        totalSpent:
          type: number
        ordersCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - loyaltyPoints
        - loyaltyTier
        - totalSpent
        - ordersCount
        - isActive
        - createdAt
...
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    AdjustLoyaltyBody:
      type: object
      properties:
        points:
          type: integer
        reason:
          type: string
      required:
        - points
        - reason
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - limit
        - totalPages

    # ── Cart ──
...
    TrackViewBody:
      type: object
      properties:
        productId:
          type: integer
        sessionId:
          type: ["string", "null"]
      required:
        - productId

    # ── Loyalty ──
...